requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

pytest==4.6.5
pytest-runner==5.1
sphinx-gherkindoc==3.6.0
behave==1.2.6

Pillow== 8.0.1
